Make documentation buildable without Git source tree (#1772)
The call to 'git rev-parse' will have a non-zero exit code when called
outside a Git working copy, aborting the documentation build. This patch
makes the documentation buildable from non-Git source trees, such as
tarballs.

This PR is part of issue #1770 to make it easier to package Open3D for
(Linux) distributions.