WebRTC callbacks
Register server callbacks triggered with messages through the WebRTC datachannel registered for specific HTML DOM elements  and events. An example of use is a radio button form in a browser that selects a tag. The browser sends a WebRTC datachannel message that triggers displaying geometry corresponding to that tag in the remote server.