Add Holes filling algorithm for CPU TSDF Pipelines.
This is as described in Section 4 of the original paper. Note: This
change has been made intentionally backwards compatible, so no
`extract_triangle_mesh` calls change the behavior.

This simple fix is usually missing on most TSDF implementations, and
depending on the application it might be a desired behavior(mostly when
reconstructing "objects" and not scenes).