Added interaction methods for Python ImageWidget
Adds and implements methods for setting MouseEvent
and KeyEvent callback functions for the ImageWidget in
Python. They are implemented identically to how this
feature is implemented for SceneWidget.

A PyImageWidget class is created similarly to PySceneWidget.
The required constructors are implemented, and the pybindings
are defined using this new class. This is only implemented for
Python, as the Mouse- and Key-methods from Widget
can already be implemented by inheriting from ImageWidget,
similarly to how this is implemented in SceneWidget.