Update CorrespondenceChecker.h
Remove incorrect doc comment

The initial version of correspondence checker was committed with a requirement that the input point clouds had the same length.  However this ceased to be required when the class was updated to only check correspondence at specified points rather than scanning all points in both inputs.  The docs were never updated with the removal of that requirement, fixing the docs here to reduce confusion.