Implement open3d::t::geometry::TriangleMesh::SelectByIndex
The method takes a list of indices and returns a new mesh built with
the selected vertices and triangles formed by these vertices.

The implementation is inspired by
  open3d::geometry::TriangleMesh::SelectByIndex.
and by
  open3d::t::geometry::TriangleMesh::SelectFacesByMask.

We first compute a mask of vertices to be selected.  If the input index
exceeds the maximum number of vertices, we throw an exception. Based on
the vertex mask we build a mapping index vector using inclusive prefix
sum algorithm. We then update the selected vertex indices in the triangles
CPU copy and build the triangles mask. I put that to a templated helper
static function to avoid the code duplication for Int32 and Int64 vertex
indices types.

With the computed masks we can select the vertices and triangles back on
the mesh's device and copy the attributes.