fix render to depth image on Apple Retina displays (#7001)
One of the few differences is that the depth version calls ConfigureForColorPicking, which disables a lot of the post processing that is used for RGB rendering, as it's not required for depth rendering. Surprisingly, it turns out that activating shadowing again for the depth rendering fixes the scaling issue and the depth image saves correctly. I was not able to find any obvious cause for this in Open3D code, so for now I am assuming that setShadowingEnabled has some unknown side effect in Filament.