Fix ICP registration pipeline documentation
The documentation says that the icp fitness is defined as the
ratio between the number of inlier correspondences and the
number of points in the target pointcloud. However, the
current implementation in Registration.cpp defines the
fitness as ratio between the number of inlier correspondences
and the number of points in the source point cloud.

This commit updates the documentation to align with the
implementation details.

Signed-off-by: Dominik Van Opdenbosch <22854909+d-vo@users.noreply.github.com>