Fix logic for adding -allow-unsupported-compiler to nvcc (#7337)
The previous logic would always add -allow-unsupported-compiler and -D_ALLOW_COMPILER_AND_STL_VERSION_MISMATCH, even when it was not needed.
The revised logic only adds the flags if MSVC>=1940 and CUDA<=12.4. The new code also uses CMAKE_CUDA_FLAGS for the define, which makes CMake's try_compile work reliably if a CMake version >=3.29.4 is used.
---------
Co-authored-by: Patrik Huber <4967343+patrikhuber@users.noreply.github.com>
Co-authored-by: Sameer Sheorey <41028320+ssheorey@users.noreply.github.com>